Remove some unused stuff
